implemented return runtime

+++ sorting_algorithms/insertion_sort.h
@@ -13,6 +13,6 @@
 
 #include "../shared/copy_array.h"
 
-int insertion_sort(short *array, int length);
+double insertion_sort(short *array, int length);
 
 #endif //TI_PROJEKT_INSERTION_SORT_H
